Added type animation to landing page
