+++ app/package.json
@@ -16,18 +16,16 @@
     "axios": "^1.3.4",
     "bcrypt": "^5.1.0",
     "bcryptjs": "^2.4.3",
-    "mongodb-memory-server": "^8.12.1",
-    "mongoose": "^7.0.3",
-    "next": "^13.2.4",
-    "node-mocks-http": "^1.12.2",
     "dotenv": "^16.0.3",
     "mongodb-memory-server": "^8.12.1",
     "mongoose": "^7.0.3",
     "next": "^13.2.4",
     "next-auth": "^4.20.1",
+    "node-mocks-http": "^1.12.2",
     "nodemailer": "^6.9.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-type-animation": "^3.0.1",
     "text-encoding": "^0.7.0",
     "usa-states": "^0.0.6",
     "yarn": "^1.22.19"
